# improves change log list view performance