--- a/changes/7613.changed
+++ b/changes/7613.changed
@@ -1,0 +1,1 @@
+Improved object change list view performance.