--- a/nautobot/ipam/tables.py
+++ b/nautobot/ipam/tables.py
@@ -1,8 +1,6 @@
-from django.db.models import Prefetch, QuerySet
+from django.db.models import Prefetch
 from django.utils.safestring import mark_safe
 import django_tables2 as tables
-from django_tables2.data import TableData
-from django_tables2.rows import BoundRows
 from django_tables2.utils import Accessor
 
 from nautobot.core.tables import (
@@ -425,17 +423,12 @@ class PrefixDetailTable(PrefixTable):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        # Conditionally prefetch children for the utilization calculation if that column is visible.
-        if self.columns["utilization"].visible and isinstance(self.data.data, QuerySet):
-            self.data = TableData.from_data(
-                self.data.data.prefetch_related(
-                    Prefetch(
-                        "children", queryset=Prefix.objects.only("network", "prefix_length", "parent_id").order_by()
-                    )
-                )
-            )
-            self.data.set_table(self)
-            self.rows = BoundRows(data=self.data, table=self, pinned_data=self.pinned_data)
+        self.add_conditional_prefetch(
+            "utilization",
+            prefetch=Prefetch(
+                "children", queryset=Prefix.objects.only("network", "prefix_length", "parent_id").order_by()
+            ),
+        )
 
     class Meta(PrefixTable.Meta):
         fields = (