# Closes DNE
# What's Changed

Adds a conditional prefetch for the `changed_object` column of the object change list view.

## Screenshots

Before:

<img width="275" height="224" alt="Screenshot 2025-07-28 at 16 00 19" src="https://github.com/user-attachments/assets/b72a06ae-b8b5-4312-97e5-2d789ca01dfd" />

After:

<img width="279" height="200" alt="Screenshot 2025-07-28 at 16 00 40" src="https://github.com/user-attachments/assets/54c89171-2737-4367-a954-9e985c5b42f6" />