@Kircheneer 2025-07-28
I am wondering at this point, does it make sense to extract the pattern used for the dynamic application of the prefetch into a function/method somewhere?