@Kircheneer 2025-07-29
> Yeah, adding a new method to `BaseTable` to encapsulate this pattern would be helpful. :-)

Done! I tried to make it somewhat generalizable but I was only able to replace one of the two other occurences of similar-ish patterns.