@Kircheneer 2025-08-05
Here is 100k:

*Without change*
<img width="211" height="272" alt="Screenshot 2025-08-05 at 09 06 33" src="https://github.com/user-attachments/assets/7f8503d2-6202-479f-8009-3d94f3980677" />

*With change*
<img width="219" height="270" alt="Screenshot 2025-08-05 at 09 07 22" src="https://github.com/user-attachments/assets/29312261-e397-42c9-bc95-c20e38b498d9" />