@Kircheneer 2025-08-05
Okay so at least on local development the actual performance impact in ms is negligible, but the amount of queries is significantly shrunk. Let me know what you think about these results