@glennmatthews 2025-08-05
What's the page size there, and does the performance change with/without the `changed_object`/`object_repr` column shown?