@Kircheneer 2025-08-06
Removing the "Object" column from the table shaves of a single query per distinct table the changed objects come from (vs. a single query per non-distinct object without the change), so that matches the expected behaviour with `prefetch_related`.

As for the page size, for >1 million change records here are two tables. Caveat - my change log data mostly contains changes for the location table, meaning that performance should be a bit worse in the *With this change* scenario in normal scenarios.

*With this change*

| Page Size | CPU Median | CPU Worst Case | DB Median | DB Worst Case |
|-|-|-|-|-|
| 50 | 170ms | 250ms | 80ms | 200ms |
| 100 | 200ms | 550ms | 80ms | 250ms |
| 250 | 600ms | 800ms | 100ms | 200ms |
| 500 | 1300ms | 1700ms | 120ms | 200ms |
| 1000 | 2500ms | 4000ms | 150ms | 250ms |

*Without this change*

| Page Size | CPU Median | CPU Worst Case | DB Median | DB Worst Case |
|-|-|-|-|-|
| 50 | 250ms | 800ms | 100ms | 300ms |
| 100 | 800ms | 1300ms | 130ms | 220ms |
| 250 | 1800ms | 2600ms | 200ms | 400ms |
| 500 | 3500ms | 5000ms | 300ms | 420ms |
| 1000 | 7000ms | 9200ms | 500ms | 680ms |

Now this is of course not really scientific, I just reloaded a bunch of times, observed, and did rough calculations, but it feels like a trend is showing.